test_generators/bls/requirements.txt
py-ecc==1.6.0
PyYAML==4.2b1
